src/server/requirements.txt
--allow-all-external
beautifulsoup4==4.4.1
Flask==0.10.1
Flask-Cors==2.1.2
iso8601==0.1.11
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
mysql==0.0.1
mysql-connector-python==2.0.4
MySQL-python==1.2.5
numpy==1.10.1
pyaml==15.8.2
pyfiglet==0.7.4
PyYAML==3.11
requests==2.8.1
six==1.10.0
Werkzeug==0.10.4
wheel==0.24.0
wikipedia==1.4.0
